+++ src/handlers/wallet/handler.rs
@@ -3,7 +3,7 @@ use super::get_calls_status::{self, GetCallsStatusError};
 use super::get_exchange_buy_status::{self, GetExchangeBuyStatusError};
 use super::get_exchange_url::{self, GetExchangeUrlError};
 use super::get_exchanges::{self, GetExchangesError};
-use super::pos::{self, BuildPosTxError, CheckPosTxError};
+use super::pos::{self, BuildPosTxsError, CheckPosTxError, SupportedNetworksError};
 use super::prepare_calls::{self, PrepareCallsError};
 use super::send_prepared_calls::{self, SendPreparedCallsError};
 use crate::error::RpcError;
@@ -96,8 +96,9 @@ pub const WALLET_GET_CALLS_STATUS: &str = "wallet_getCallsStatus";
 pub const PAY_GET_EXCHANGES: &str = "reown_getExchanges";
 pub const PAY_GET_EXCHANGE_URL: &str = "reown_getExchangePayUrl";
 pub const PAY_GET_EXCHANGE_BUY_STATUS: &str = "reown_getExchangeBuyStatus";
-pub const POS_BUILD_TRANSACTION: &str = "reown_pos_buildTransaction";
-pub const POS_CHECK_TRANSACTION: &str = "reown_pos_checkTransaction";
+pub const POS_BUILD_TRANSACTIONS: &str = "wc_pos_buildTransactions";
+pub const POS_CHECK_TRANSACTION: &str = "wc_pos_checkTransaction";
+pub const POS_SUPPORTED_NETWORKS: &str = "wc_pos_supportedNetworks";
 
 #[derive(Debug, Error)]
 enum Error {
@@ -125,11 +126,14 @@ enum Error {
     #[error("{PAY_GET_EXCHANGE_BUY_STATUS}: {0}")]
     GetExchangeBuyStatus(GetExchangeBuyStatusError),
 
-    #[error("{POS_BUILD_TRANSACTION}: {0}")]
-    PosBuildTransaction(BuildPosTxError),
+    #[error("{POS_BUILD_TRANSACTIONS}: {0}")]
+    PosBuildTransactions(BuildPosTxsError),
 
     #[error("{POS_CHECK_TRANSACTION}: {0}")]
     PosCheckTransaction(CheckPosTxError),
+
+    #[error("{POS_SUPPORTED_NETWORKS}: {0}")]
+    PosSupportedNetworks(SupportedNetworksError),
 
     #[error("Method not found")]
     MethodNotFound,
@@ -158,8 +162,9 @@ impl Error {
             Error::GetExchanges(_) => -6,
             Error::GetUrl(_) => -7,
             Error::GetExchangeBuyStatus(_) => -8,
-            Error::PosBuildTransaction(_) => -9,
+            Error::PosBuildTransactions(_) => -9,
             Error::PosCheckTransaction(_) => -10,
+            Error::PosSupportedNetworks(_) => -11,
             Error::MethodNotFound => -32601,
             Error::InvalidParams(_) => -32602,
             Error::Internal(_) => -32000,
@@ -176,8 +181,9 @@ impl Error {
             Error::GetExchanges(e) => e.is_internal(),
             Error::GetUrl(e) => e.is_internal(),
             Error::GetExchangeBuyStatus(e) => e.is_internal(),
-            Error::PosBuildTransaction(e) => e.is_internal(),
+            Error::PosBuildTransactions(e) => e.is_internal(),
             Error::PosCheckTransaction(e) => e.is_internal(),
+            Error::PosSupportedNetworks(e) => e.is_internal(),
             Error::MethodNotFound => false,
             Error::InvalidParams(_) => false,
             Error::Internal(_) => true,
@@ -300,14 +306,14 @@ async fn handle_rpc(
             .map_err(Error::GetExchangeBuyStatus)?,
         )
         .map_err(|e| Error::Internal(InternalError::SerializeResponse(e))),
-        POS_BUILD_TRANSACTION => serde_json::to_value(
-            &pos::build_transaction::handler(
+        POS_BUILD_TRANSACTIONS => serde_json::to_value(
+            &pos::build_transactions::handler(
                 state,
                 project_id,
                 serde_json::from_value(params).map_err(Error::InvalidParams)?,
             )
             .await
-            .map_err(Error::PosBuildTransaction)?,
+            .map_err(Error::PosBuildTransactions)?,
         )
         .map_err(|e| Error::Internal(InternalError::SerializeResponse(e))),
         POS_CHECK_TRANSACTION => serde_json::to_value(
@@ -318,6 +324,12 @@ async fn handle_rpc(
             )
             .await
             .map_err(Error::PosCheckTransaction)?,
+        )
+        .map_err(|e| Error::Internal(InternalError::SerializeResponse(e))),
+        POS_SUPPORTED_NETWORKS => serde_json::to_value(
+            &pos::supported_networks::handler(state, project_id)
+                .await
+                .map_err(Error::PosSupportedNetworks)?,
         )
         .map_err(|e| Error::Internal(InternalError::SerializeResponse(e))),
         _ => Err(Error::MethodNotFound),